move some files (entirely or partially) from alias-classes to effects-util to avoid library dependences with effects librairies.
This was more work than first expected  :-( .

Also remove many #include "alias-classes.h" and #include "effects-util.h" when possible.

Library dependences between effects and alias-classes are now :

effects-util -> effects-generic -> effects-[simple-convex] -> alias-classes

Some more work to compile with strict options with autotools (thanks to SG).



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@17426 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/ri-util/constant.c
@@ -400,7 +400,7 @@ entity float_to_entity(float c)
 entity int_to_entity(_int c)
 {
     string num;
-    asprintf(&num, "%d", c);
+    asprintf(&num, "%d", (int) c);
     entity e = MakeConstant(num,is_basic_int);
     free(num);
     return e;
